some update

+++ src/coords.hpp
@@ -5,6 +5,6 @@ struct Coords {
         return std::sqrt(pow(rhs.x - this->x, 2) + pow(rhs.y - this->y, 2));
     }
 
-    size_t x;
-    size_t y;
+    float x;
+    float y;
 };
